move `ChainId` definition from messenger to domains primitives
