+++ crates/subspace-malicious-operator/src/bin/subspace-malicious-operator.rs
@@ -27,8 +27,7 @@ use sc_utils::mpsc::tracing_unbounded;
 use serde_json::Value;
 use sp_core::crypto::Ss58AddressFormat;
 use sp_core::traits::SpawnEssentialNamed;
-use sp_domains::DomainId;
-use sp_messenger::messages::ChainId;
+use sp_domains::{ChainId, DomainId};
 use std::collections::HashMap;
 use subspace_malicious_operator::malicious_domain_instance_starter::DomainInstanceStarter;
 use subspace_malicious_operator::{Cli, DomainCli};
