+++ domains/client/relayer/src/worker.rs
@@ -5,8 +5,7 @@ use sc_client_api::{AuxStore, BlockchainEvents, ProofProvider};
 use sc_state_db::PruningMode;
 use sp_api::{ApiError, ProvideRuntimeApi};
 use sp_consensus::SyncOracle;
-use sp_domains::DomainsApi;
-use sp_messenger::messages::ChainId;
+use sp_domains::{ChainId, DomainsApi};
 use sp_messenger::{MessengerApi, RelayerApi};
 use sp_runtime::scale_info::TypeInfo;
 use sp_runtime::traits::{CheckedSub, NumberFor, Zero};
