+++ domains/pallets/messenger/src/fees.rs
@@ -3,8 +3,9 @@ use crate::{BalanceOf, Config, Error, Pallet};
 use frame_support::traits::fungible::Mutate;
 use frame_support::traits::tokens::{Fortitude, Precision};
 use frame_support::weights::WeightToFee;
+use sp_domains::ChainId;
 use sp_messenger::endpoint::Endpoint;
-use sp_messenger::messages::{ChainId, ChannelId, FeeModel, MessageId, Nonce};
+use sp_messenger::messages::{ChannelId, FeeModel, MessageId, Nonce};
 use sp_messenger::OnXDMRewards;
 use sp_runtime::traits::CheckedAdd;
 use sp_runtime::DispatchResult;
