+++ domains/primitives/messenger-host-functions/src/host_functions.rs
@@ -5,8 +5,7 @@ use sp_api::ProvideRuntimeApi;
 use sp_blockchain::HeaderBackend;
 use sp_core::traits::CodeExecutor;
 use sp_core::H256;
-use sp_domains::{DomainId, DomainsApi};
-use sp_messenger::messages::ChainId;
+use sp_domains::{ChainId, DomainId, DomainsApi};
 use sp_messenger::MessengerApi;
 use sp_runtime::traits::{Block as BlockT, Header, NumberFor};
 use std::marker::PhantomData;
