+++ domains/service/src/domain.rs
@@ -31,9 +31,8 @@ use sp_consensus::SyncOracle;
 use sp_consensus_slots::Slot;
 use sp_core::traits::SpawnEssentialNamed;
 use sp_core::{Decode, Encode, H256};
-use sp_domains::{BundleProducerElectionApi, DomainId, DomainsApi, OperatorId};
+use sp_domains::{BundleProducerElectionApi, ChainId, DomainId, DomainsApi, OperatorId};
 use sp_domains_fraud_proof::FraudProofApi;
-use sp_messenger::messages::ChainId;
 use sp_messenger::{MessengerApi, RelayerApi};
 use sp_mmr_primitives::MmrApi;
 use sp_offchain::OffchainWorkerApi;
