+++ domains/test/primitives/Cargo.toml
@@ -12,8 +12,9 @@ include = [
 ]
 
 [dependencies]
-codec = { package = "parity-scale-codec", version = "3.4.0", default-features = false, features = ["derive"]}
+codec = { package = "parity-scale-codec", version = "3.4.0", default-features = false, features = ["derive"] }
 sp-api = { version = "4.0.0-dev", git = "https://github.com/subspace/polkadot-sdk", rev = "d6b500960579d73c43fc4ef550b703acfa61c4c8", default-features = false }
+sp-domains = { version = "0.1.0", default-features = false, path = "../../../crates/sp-domains" }
 sp-messenger = { version = "0.1.0", default-features = false, path = "../../primitives/messenger" }
 subspace-runtime-primitives = { version = "0.1.0", path = "../../../crates/subspace-runtime-primitives", default-features = false }
 
@@ -21,6 +22,7 @@ subspace-runtime-primitives = { version = "0.1.0", path = "../../../crates/subsp
 default = ["std"]
 std = [
     "sp-api/std",
+    "sp-domains/std",
     "sp-messenger/std",
     "subspace-runtime-primitives/std",
 ]
