+++ domains/test/service/src/domain.rs
@@ -29,8 +29,7 @@ use serde::de::DeserializeOwned;
 use sp_api::{ApiExt, ConstructRuntimeApi, Metadata, ProvideRuntimeApi};
 use sp_block_builder::BlockBuilder;
 use sp_core::{Decode, Encode, H256};
-use sp_domains::DomainId;
-use sp_messenger::messages::ChainId;
+use sp_domains::{ChainId, DomainId};
 use sp_messenger::{MessengerApi, RelayerApi};
 use sp_offchain::OffchainWorkerApi;
 use sp_runtime::traits::{Dispatchable, NumberFor};
